+++ vscode-test-runner/workspace-1/.joyride/scripts/workspace_activate.cljs
@@ -1,7 +1,6 @@
 (ns workspace-activate
   (:require [integration-test.db :as db]
-            ["vscode" :as vscode]
-            [promesa.core :as p]))
+            ["vscode" :as vscode]))
 
 #_{:clj-kondo/ignore [:clojure-lsp/unused-public-var]}
 (def symbol-1 :symbol-1)
@@ -14,14 +13,4 @@
 (defn ws-root []
   (first vscode/workspace.workspaceFolders))
 
-;; Wait for the waiter promise to be created and then immediately resolve it
-(defn- activated! []
-  (if-let [waiter (:ws-activate-waiter @db/!state)]
-    (do
-      (println "Runner: workspace-activate script done, resolving waiter promise")
-      (p/resolve! waiter))
-    (do
-      (println "Runner: workspace-activate script done, but no waiter promise created yet")
-      (js/setTimeout activated! 10))))
-
-(activated!)
+(swap! db/!state assoc :ws-activated? true)