+++ vscode-test-runner/workspace-1/.joyride/src/subdir/require_js_from_subdir.cljs
@@ -1,0 +1,14 @@
+(ns subdir.require-js-from-subdir
+  (:require ["../js-file.js" :as js-file]))
+
+(defn hello []
+  (js-file/hello))
+
+(def fortytwo js-file/fortytwo)
+
+(comment
+  (hello)
+  fortytwo
+  :rcf)
+
+fortytwo