+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes to Joyride
 
 - Fix [Error `No context found` when evaluating `is` form](https://github.com/BetterThanTomorrow/joyride/issues/124)
 - [Add e2e testing](https://github.com/BetterThanTomorrow/joyride/issues/125)
+- Fix [Requiring a script that require a js file fails](https://github.com/BetterThanTomorrow/joyride/issues/128)
+- Fix [Requiring npm modules by proxy fail](https://github.com/BetterThanTomorrow/joyride/issues/129)
 
 ## [0.0.28] - 2022-12-21